Rework the supplementary type indicator array ownership and discovery API
The array and file based datasources have an optional supplementary
type indicator array.

Previously, the datasources used the passed-in array pointer as is,
making it necessary to keep the array alive during the lifetime of
the report.

The type indicator array ownership reworked, so the array and file
based datasources make a copy of the type indicator array.

The array discovery API was modified: ocrpt_query_discover_array()
has a new boolean output parameter to indicate that the returned
array is allocated, so it must be freed after use.

The C API always returns "false" for freeing the type indicator
array, leaving it under application control.

The PHP module was adapted to this change and returns true if
the PHP array was successfully converted to a C array.

The C unit test array_resolve_test.c was adapted to the API change.

No change in unit test results.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>