Extend a C unit test
Extend expr_format_test.c with an expression using the format()
function to prove that an obscure format string flag to truncate
a string to n characters (i.e.: "%9.9s") works for UTF-8 characters.

Also extend the same unit test with the exact formatting example
from the documentation. It was already there, but without the
string continuation quirk.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>