Introduce the unary plus function
Added an uplus() function and modified the grammar to use
this function when the unary '+' operator is used.

It is a NOP or pass-through for numeric operands and is
eliminated by the expression optimizer for numeric constants.

It catches bad expressions that use the unary plus operator
for types other than numeric.

The C and PHP unit test results for grammar_test.c and
grammar_test.php was changed.

Extended the C unit test called expr_test.c to exercise
the unary plus in a minimal way.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>