Make it conditional to emit usual report elements
This will allow fanout of data rows for <Graph> and <Chart>.
Added TODO lines where it should be done.

The plan is that the tabular elements are primary and only
output <Graph> and <Chart> if they are the only element
present in a <Report>.

No change in unit test results.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>