Conditionally use static array backed PHP array datasource
It turned out that with older PHP versions, unit tests using
PHP array queries fail with PHP 7.2 and older after switching to
the dynamic array handling in PHP module.

Restore the old method to use a C array converted from the
PHP array data and make it conditional for PHP 7.3 and newer.
The old method can be used with the configure option
--with-opencreports-static-array.

This old method is now similarly handled as the file based
datasources and does not need global accounting of the backing
arrays. Instead, the backing array is internal to the PHP array
query data. This can lead to higher memory usage if multiple
array queries use the same PHP array, but safer.

The C array backing method is unconditionally used for PHP 7.2
and older.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>