More backward compatibility changes
Allow building without PostgreSQL support if PQdescribePortal()
is not present. This was introduced in PostgreSQL 8.2.0 in 2006.

Allow to disable building the documentation with the new
--enable-documentation configure option. Building documentation
is still gated by the presence of the necessary commands.

The C compiler option -Werror is not added when the configure
option --enable-maintainer-mode is in effect.

Added #include <unistd.h> to opencreport.h so the definition
of __restrict is valid.

Fixed the compiler error in db-source.c with GCC 4.4 in the
designated initialized for database sources' connect parameter
lists. Anonymous unions need the extra brackets in the
initializer.

Don't typedef ocrpt_color twice. Removed the one from output.h
because it is already present in the public opencreport.h header.
GCC 4.4 complained about this.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>