Replace home-grown AX_TRY_LINK_FLAG_CXX with AC_LIBTOOL_LINKER_OPTION
AC_LIBTOOL_LINKER_OPTION is a compatibility alias for
_LT_LINKER_OPTION, which was implicitly used in AX_TRY_LINK_FLAG_CXX
with an added message and an incorrectly implemented extra linking
step.

When trying to run ./configure the first time after git clone, this
extra linking step failed with "./libtool: No such file or directory".

Use AC_LIBTOOL_LINKER_OPTION directly which results in the same
linker option to be detected and set for the unit tests.
It is done correctly at the first time, too.

The only downside is losing the output message about which linker
option is being used.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>