Modify return type of layout detail API functions
Changed return type of layout detail API functions that
internally create expressions from void to ocrpt_expr *,
so the internal expression is exposed.

It is useful for probing / debugging internal expressions that
may be derived from datasource row values in unit tests or
in production via row level callbacks.

Adapted the PHP binding and the documentation to the changes.

TODO: add the getter part of the API so such layout detail
expressions created internally via a report XML descriptor
can also be probed.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>