Prevent modifying report structure and expressions while executing
Added an "executing" flag to the main opencreport structure
and set and unset it for executing reports parts.

With the help of this flag, every API that adds output items
(line, horizontal line, text, image and barcode) or sets their
detail expressions now refuse to change internal data when
the report is executing.

This will allow adding an API to iterate over output details
that may be used safely in report callbacks.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>