Add -Wno-c23-extensions to CFLAGS before detecting the dynamic linker option
CLANG throws this when detecting which variant of the
"export dynamic" linker options is supported:

| warning: omitting the parameter name in a function definition
| is a C23 extension [-Wc23-extensions]

This is due to the boilerplate header used by the
AC_LIBTOOL_LINKER_OPTION autoconf macro:

| lt_simple_link_test_code='int main(int, char *[[]]) { return(0); }'

AC_LIBTOOL_LINKER_OPTION does not check whether the warning
reports the linker option not being supported. As a result,
none of the linker option variants are indicated as supported,
which makes the configure script fail.

Add -Wno-c23-extensions to CFLAGS temporarily to prevent this.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>