Extend ignore.supp to cover libpython
Py_Initialize() and friends, and PyImport_Import*() calls allocate
memory for the lifetime of the application without Py_Finalize()
freeing it.

This allocation is static, so strictly speaking it's not a leak.
Still, Valgrind and AddressSanitizer complain about it.

Add two entries to ignore.supp to cover this, so unit tests won't
show errors due to memory leaks.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>