Add arraydescr_richcompare function to ignore.supp
pandas.ExcelFile.parse() uses NumPy behind the scenes.

The Pandas DataFrame seems to have a static global allocation
created by NumPy, which does not get freed if "del dataframe"
is executed in Python or if the reference on the dataframe
object structure is released.

Furthermore, this allocation does not get repeated when
pandas.ExcelFile.parse() is called multiple times.

The first C function in the AddressSanitizer leak trace is
arraydescr_richcompare. Add this function to ignore.supp to
suppress this static allocation.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>