Protect ocrpt_postgresql_describe_base() against too many columns
It isn't a problem in practice, but some dumb static analyzers
report a potential integer overflow by using "unsanitized input
from a database", with the index value reaching the number of
columns multiplied by 3.

No database supports 715 million columns in a query to overflow
INT_MAX. Add a safety check in ocrpt_postgresql_describe_base()
for this detail, plus an assert() call for static analyzers that
can't deduce the detail from the check.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>